backend/ai_service/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
torch
transformers
sentencepiece
pydantic
datasets
